refactor : 클래스명 및 패키지 변경, 응답방식 수정에 따른 코드 리팩토링
- Json형태로 응답을 위해 코드 수정
