+++ src/main/java/com/zerototen/savegame/config/WebConfig.java
@@ -1,6 +1,6 @@
 package com.zerototen.savegame.config;
 
-import com.zerototen.savegame.converter.StringToLocalDateConverter;
+import com.zerototen.savegame.util.StringToLocalDateConverter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
